bid: tyding up - desplit files (now use 3 for bid) - now include only ctypes for the kernel - rename bid_init and bid_deinit - better code format - changed error line detection - [other]
