--- sources/libraries/ansi/ctype.h
@@ -1,33 +1,0 @@
-#pragma once
-
-/* --- 7.4 Character - handling --------------------------------------------- */
-
-/* --- 7.4.1 Character classification - Copying functions ------------------- */
-
-int isalnum(int c);
-
-int isalpha(int c);
-
-int isblank(int c);
-
-int iscntrl(int c);
-
-int isdigit(int c);
-
-int isgraph(int c);
-
-int islower(int c);
-
-int isprint(int c);
-
-int ispunct(int c);
-
-int isspace(int c);
-
-int isupper(int c);
-
-int isxdigit(int c);
-
-int tolower(int c);
-
-int toupper(int c);
