+++ sources/libraries/bid/bid.h
@@ -40,21 +40,12 @@ struct bid
     int _current_scanned_token_cursor;
 };
 
-typedef Result(struct bid_error, MonoState) BidParseResult;
 typedef Result(struct bid_error, struct bid) BidResult;
 
-BidResult init_bid(Str idl_in);
+BidResult bid_init(Str idl_in);
 
-void destroy_bid(struct bid *in);
-
-enum bid_token_type bid_token_from_char(char from);
-
-enum bid_keywords bid_keyword_from_string(Str from);
+void bid_deinit(struct bid *in);
 
 struct bid_error bid_create_error(enum bid_error_type type, Str error_msg, const struct bid *idl_in);
 
 struct bid_error bid_create_unexpected_token_error(Str expected_token, const struct bid *idl_in);
-
-int bid_is_keyword(int chr);
-
-void skip_comment_and_space(struct bid *idl_in);
