--- sources/libraries/bid/c_convert/bid_convert_type.h
@@ -1,7 +1,0 @@
-#pragma once
-#include <brutal/base.h>
-#include <bid/bid.h>
-#include <bid/parsing/bid_ast.h>
-#include <brutal/io.h>
-
-void bid_write_type(IoWriter *writer, struct bid_ast_node *type);
