+++ sources/libraries/bid/generator-c.h
@@ -1,7 +1,7 @@
 #pragma once
-#include <brutal/base.h>
+#include <bid/ast.h>
 #include <bid/bid.h>
-#include <bid/parsing/bid_ast.h>
+#include <brutal/base.h>
 #include <brutal/io.h>
 
 void convert_bid_to_c(const struct bid *bid, IoWriter *writer);
