--- sources/libraries/bid/parsing/bid_error.h
@@ -1,5 +1,0 @@
-#pragma once
-#include <bid/parsing/bid_ast.h>
-#include <brutal/base.h>
-
-BidParseResult scan_error(struct bid *idl_in, struct bid_ast_node *method_node);
